+++ src/test/engine-integrations/full-margin/FullMarginFixture.t.sol
@@ -38,10 +38,10 @@ abstract contract FullMarginFixture is Test, ActionHelper, Utilities {
     address internal bob;
 
     // usdc collateralized call / put
-    uint32 internal pidUsdcCollat;
+    uint40 internal pidUsdcCollat;
 
     // eth collateralized call / put
-    uint32 internal pidEthCollat;
+    uint40 internal pidEthCollat;
 
     uint8 internal usdcId;
     uint8 internal wethId;
@@ -103,7 +103,7 @@ abstract contract FullMarginFixture is Test, ActionHelper, Utilities {
     function mintOptionFor(
         address _recipient,
         uint256 _tokenId,
-        uint32 _productId,
+        uint40 _productId,
         uint256 _amount
     ) internal {
         address anon = address(0x42424242);
